+++ src/sagas/getGenres.js
@@ -1,0 +1,16 @@
+import axios from "axios";
+import { put } from "redux-saga/effects";
+
+/**
+ * Gets all the genres from the database
+ */
+function* getGenres() {
+  try {
+    const response = yield axios.get("/api/movies/");
+    yield put({ type: "SET_GENRES", payload: response });
+  } catch (err) {
+    yield console.warn("err @getGenres*:", err);
+  }
+}
+
+export default getGenres;
